pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agentic-onebotads"
version = "0.1.0"
description = "Local-first agentic ad generation backend for Agentic OneBotAds."
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
  "chromadb>=1.0.0",
  "diffusers>=0.32.0",
  "fastapi>=0.115.0",
  "httpx>=0.27.0",
  "accelerate>=1.2.0",
  "langchain>=0.3.0",
  "langchain-community>=0.3.0",
  "langchain-ollama>=0.2.0",
  "llama-index-core>=0.12.0",
  "llama-index-embeddings-ollama>=0.5.0",
  "llama-index-llms-ollama>=0.5.0",
  "llama-index-readers-file>=0.4.0",
  "llama-index-vector-stores-chroma>=0.4.0",
  "pandas>=2.2.0",
  "pydantic-settings>=2.6.0",
  "python-dotenv>=1.0.0",
  "python-multipart>=0.0.9",
  "safetensors>=0.4.5",
  "torch>=2.5.0",
  "transformers>=4.47.0",
  "uvicorn[standard]>=0.30.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "pytest-asyncio>=0.24.0",
  "pytest-cov>=5.0.0",
  "ruff>=0.8.0",
]

[tool.setuptools]
package-dir = {"" = "apps/api/src"}

[tool.setuptools.packages.find]
where = ["apps/api/src"]
include = ["onebot_ads*"]

[tool.pytest.ini_options]
addopts = "-q"
pythonpath = ["apps/api/src"]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"
src = ["apps/api/src", "tests"]

[tool.ruff.lint]
select = ["B", "E", "F", "I"]
